refactor: delete `OpenRecentMenuViewModelJavaFxExtensionsTest`
I hit a roadblock after adding the `SeparatorMenuItem` in the mix (it wouldn't instantiate without the JavaFx application thread). For these tests to be complete and accurate, we need to run these tests on the JavaFx application thread. I tried to run these tests using a JavaFx test runner and the experiment was successful, but it fell short in two places.

1) The tests are too slow.
2) They fail on the CI.

For these reasons, we won't test this logic via automated tests and rely on manual testing instead.